+++ src/Plc/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ___
 
+## 1.4.0 (2020-04-20)
+
+### Changed
+
+- Log messages are now a little bit more detailed.
+- The classes ***Plc***, ***PlcItem***, ***BitCollection*** implement the **IFormattable** interface as to format their string representation for output purposes.
+___
+
 ## 1.3.0 (2020-04-07)
 
 ### Added